restore ArgumentTypeWrapper.byName() and ArgumentTypeWrapper.printAll()
